Sonar fixes
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1705587 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/ddf/EscherDump.java
@@ -744,7 +744,7 @@ public final class EscherDump {
         String result = "";
         result += (short) ( n32 >> 16 );
         result += '.';
-        result += (short) ( n32 & (short) 0xFFFF );
+        result += (short) ( n32 & 0xFFFF );
         return result;
     }
